Separate symbols
