+++ Cargo.toml
@@ -1,3 +1,4 @@
 [workspace]
-members = ["rusl", "tiny-std"]
+resolver = "2"
+members = ["rusl", "tiny-start", "tiny-std"]
 exclude = ["test-runners/alloc-st-main", "test-runners/no-alloc-main", "test-runners/threaded-main", "test-runners/test-lib"]
