+++ tiny-start/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "tiny-start"
+version = "0.1.0"
+edition = "2021"
+license = "MPL-2.0"
+readme = "../Readme.md"
+repository = "https://github.com/MarcusGrass/tiny-std"
+homepage = "https://github.com/MarcusGrass/tiny-std"
+description = "Tiny start functions and symbol relocation"
+documentation = "https://docs.rs/tiny-std"
+categories = ["external-ffi-bindings", "no-std", "os"]
+keywords = ["ffi", "bindings", "operating", "system", "linux"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[features]
+default = []
+aux = ["start"]
+start = []
+mem-symbols = []
+
+
+[dependencies]
+rusl = { path = "../rusl", default-features = false }
