+++ tiny-start/src/elf.rs
@@ -1,0 +1,3 @@
+pub mod aux;
+
+pub(crate) mod dynlink;
