+++ tiny-start/src/symbols.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "mem-symbols")]
+pub mod mem;
